[US_ND] Hydrate state_staff_location_periods (Recidiviz/recidiviz-data#19653)
## Description of the change

Hydrates the StateStaffLocationPeriod entity for supervision officers
only in ND. Hydrated from docstars_officers raw data. This is the second
step in addressing the issue of inactive officers appearing in Vitals in
ND (https://github.com/Recidiviz/recidiviz-data/issues/19264).

The `docstars_officers` table that we receive is always a full
historical export, and has a field called `STATUS` to mark whether or
not officers are active or not. This view uses that field to close
periods that end in termination of an officer's employment. There should
be no instances of officers having an open period while their `STATUS =
0`.

When an officer becomes inactive, their last location period has an end
date, and there is no subsequent open period for that officer. The start
and end dates of location periods are hydrated from the `RecDate` field,
which corresponds to the date the record was created in the Docstars
system.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#19264

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 0f2479ecc6092d50dfa2edd9443e0bb7147051de