Updates assessment_score_sessions (Recidiviz/recidiviz-data#19815)
## Description of the change

This PR updates `assessment_score_sessions` to allow multiple
assessments of different types to fall on the same `assessment_date`.
This could happen if, for example, a client has both a general and
specialty assessment performed on the same day.

I also added a WHERE clause to remove rows where there is no score or
assessment level produced by the assessment (a rare but not impossible
occurrence). Finally, I added some details to the view description and
removed the indent from the query to improve readability in GBQ.

These changes reduced the row count by ~1,000 (of >3.3M), so we
shouldn't expect substantial changes in downstream metrics.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes None, related to Recidiviz/recidiviz-data#19723

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8df50c0e286e513fdb8422485b874a523c0591d5