[Workflows Remove all staff from staff_record that do not have caseloads (Recidiviz/recidiviz-data#19773)
## Description of the change

Removes all staff from the `staff_record` who do not have either a
supervision or facility caseload.

This has been run in the jovergaag_19771_* sandbox in both staging and
production

The comparison dataset is
`jovergaag_19771_compare_workflows_views_comparison_output` in both
staging and production. As far as I can tell, the only differences are
the staff where `has_caseload` but would love a second pair of eyes on
it.

The number of staff in the new staff_record is now exactly the number as
```
select count(*)
from recidiviz-123.workflows_views.staff_record_materialized
where (has_caseload is not false or has_facility_caseload is not false)
```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#19771

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 402200586d5426d5f6c61ef2ea4362a136e5ea24