[Workflows] Add look-back for populated officer id for US_MI in client_record (Recidiviz/recidiviz-data#19842)
## Description of the change

Most open supervision officer spans are missing an officer id in MI due
to a known ingest bug. In the meantime, we'll add a bypass to use the
most recent officer assigned to a person in MI instead of whatever is on
the open period.

I've filed Recidiviz/recidiviz-data#19840 to remove this once the supervision periods are fixed.

I exported the new version of the client record to my scratch folder
(`recidiviz-staging.kesich_scratch.client_test`) and checked that the
exported clients were identical in every state except MI (and ME since
they were just enabled there as well since the last deploy).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- ~[ ] Tests have been written to cover the code changed/added as part
of this pull request~ See Above

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b0dae07fbacf67ba71ac1779355a064226625fc8