[US_ND] Hydrate state_staff_role_period (Recidiviz/recidiviz-data#19809)
## Description of the change

Hydrates the StateStaffRolePeriod entity for supervision officers in ND
using information from `docstars_officers`.

The `docstars_officers` table that we receive is always a full
historical export, and has a field called `STATUS` to mark whether or
not officers are active or not. This view uses that field to close
periods of an officer having a given role. Since the only role currently
included in this view is `StateStaffRoleType.SUPERVISION_OFFICER`, these
periods are proxies for an officer's employment. There should be no
instances of officers having an open period while their `STATUS = 0.`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#19807

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fe0fb206057987a328fafbbbb4d283f37694d338