[Workflows][US_MI] Make case_notes an array for CR record (Recidiviz/recidiviz-data#19926)
## Description of the change

The ETL delegate is failing for every row of
`us_mi_complete_classification_review_form_record.json` in staging. The
delegate expects `case_notes` to be an array, not a structure. This PR
changes the column to be an array with a single entry instead of a bare
object.

I've run this manually, saved the results to
`recidiviz-staging.kesich_scratch.
us_mi_complete_classification_review_form_record_materialized`, and
exported the results to the ETL bucket. The delegate ran correctly and
populated the `US_MI-classificationReviewReferrals` collection in
staging.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- ~[ ] Tests have been written to cover the code changed/added as part
of this pull request~ BQ change. See testing above

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 631ac2c9d8f2d081b0fb7f250801742cee9c0aeb