[Dataflow] Bugfixes / improvements related to running sandbox with person_filter_ids (Recidiviz/recidiviz-data#19915)
## Description of the change

Two fixes I made to allow me to run a sandbox pipeline that only
filtered down to process data for a single person:
1) Fixed the argument formatting so that the list of person_ids would be
converted to a list to be passed to the template (like we do for
metric_types)
2) Removed the logic where we terminate the pipeline before writing any
rows when the person_filter_ids flag is set. I am pretty sure this logic
is a legacy relic of a time when we would run sandbox pipelines that
wrote out to the normal staging datasets and didn't have a way to output
to a personal sandbox. Now that the sandbox pipeline script routes
output elsewhere, there's no reason why we shouldn't allow you to output
filtered results.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Fixed while investigating Recidiviz/recidiviz-data#19912

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 254432ee1feb05c7227d1b0c7c3a5511ab599fb1