[Justice Counts] save annual data when uploaded for monthly metrics (Recidiviz/recidiviz-data#19936)
## Description of the change

Previously, we were not saving data if annual data was provided via bulk
upload for a metric configured with a monthly reporting frequency. This
PR changes this behavior so that now we save the data, and warn the user
/ direct them to change the reporting frequency on the Metric Config
page.

Warnings when
1. Annual data is uploaded for metric configured as monthly
2. Monthly data is uploaded for metric configured as annual
<img width="767" alt="Screenshot 2023-04-03 at 2 42 03 PM"
src="https://user-images.githubusercontent.com/82831800/229610764-7c544b5c-12aa-480a-84b6-388d9d4d247e.png">

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#19481
Closes Recidiviz/recidiviz-data#19678

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9f71551c3f199c614bd7ffb7823db6719925a0c8