[Justice Counts] Remove JC usages of EntityEnum (Recidiviz/recidiviz-data#19947)
## Description of the change

Eliminates the `Dimension` class which is unused now that
manual_upload.py is deleted. Updates all usages of `Dimension` to
`DimensionBase` and also removes `EntityEnum`/`EntityEnuMeta` references
from JC code which are also no longer needed now that the functionality
in `Dimension` is gone.

It its totally possible that all of these dimension classes can be
deleted entirely, but that felt like a bigger lift and this is the
minimal refactor that unblocks me.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Unblocks cleanup for Recidiviz/recidiviz-data#8905

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ae3791feb19bc5edced3c86d1aa61848d4677c40