[US_IX] Fix supervision contacts duplicate external IDs (Recidiviz/recidiviz-data#20020)
## Description of the change

In [this PR](https://github.com/Recidiviz/recidiviz-data/pull/19811), I
revised the date filters for supervision contacts so that it pulled both
legacy and new contacts for the date 11/10/2022 (since testing with POs
indicated that there were both legacy contacts and new contacts for that
date). While the query made sure that the "legacy contacts" subquery
only pulled in legacy contacts, it didn't prevent the "new contacts"
subquery from accidentally pulling in legacy contacts from that day.
This led to a duplicate external id issue that's currently crashing the
partial rerun.

This PR fixes the "new contacts" subquery to make sure it doesn't pull
in any contacts that have already been counted as a legacy contact.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4ddbf8e47a971200b5f67b1140c7ba2188c05188