[US_MI] Remove supervision level mapping for interstate compact (Recidiviz/recidiviz-data#20030)
## Description of the change

Remove supervision level mapping for interstate compact so that we can
preserve supervision level even for IC cases.

In [this PR](https://github.com/Recidiviz/recidiviz-data/pull/19672) we
added in a mapping condition for supervision levels such that
supervision periods with active interstate compact legal orders get
mapped to `StateSupervisionLevel.INTERSTATE_COMPACT `. We are now
removing that mapping condition because we realized we actually want to
preserve the supervision level even if it's an interstate compact case
(since IC cases will still be included for opportunities).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d041f04397e48b9997dea7951c57108619d68497