Supports more flexible aggregation level object definition (Recidiviz/recidiviz-data#19941)
## Description of the change

This PR makes the following changes:
- allows users to specify a source query as input to a
`MetricAggregationLevel`, where previously the input required the name
of a table in the `sessions` dataset. This allows us to support more
flexible disaggregations or attributes that may not be present in
off-the-shelf views
- includes separate parameters for the `primary keys` vs. `attributes`
associated with an aggregation level. These lists are combined in most
places to return the full list of index columns to be included in any
derivative queries, but the separation provides a few benefits: 1)
allows for more visibility into the actual primary keys of an
aggregation level, 2) we can still add auxiliary attribute columns to
our tables while retaining the ability to use primary keys on their own
for joins with other tables that might not contain auxiliary attributes
(e.g., most officer tables have `officer_id`, but most don't have
`officer_name`) and also in contexts where attributes may be duplicative
(e.g., `clustering_fields` only needs to cluster on primary keys, and
will in fact throw an error if too many columns are included)
- adds the "pretty" display names for all locations and officers as
`attribute_columns`, allowing us to observe displayable names directly
within the metric tables rather than having to join them in further
downstream.
- as part of the change above, brings `facility_name` through
`location_sessions`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#19901
Closes Recidiviz/recidiviz-data#19664

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 36783b67d9a405cf4b4a28e47edbfbe01ea5b3ee